플롯이 아니라 더블을 넘기는 토탈 세컨트 확장

+++ VKLib/NativeExtension/NumberExtensions.cs
@@ -135,12 +135,13 @@ namespace VKLib.NativeExtension
 
     public static class DateTimeExtension
     {
-        public static float GetTotalSeconds(this DateTime time)
+        public static double GetTotalSeconds(this DateTime time)
         {
             var dtEnd = new DateTime();
             var elapsed = time - dtEnd;
 
-            return (float) elapsed.TotalSeconds;
+            return (double) elapsed.TotalSeconds;
         }
+        
     }
 }
